requirements.txt
discord.py>=2.2.0
python-dotenv>=0.21.0

